+++ query.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"database/sql"
 	"fmt"
+	"reflect"
 	"time"
 
 	"github.com/pkg/errors"
@@ -287,7 +288,7 @@ func (rows *Rows) Values() ([]interface{}, error) {
 		}
 
 		if dt, ok := rows.conn.ConnInfo.DataTypeForOID(fd.DataType); ok {
-			value := dt.Value
+			value := reflect.New(reflect.ValueOf(dt.Value).Elem().Type()).Interface().(pgtype.Value)
 
 			switch fd.FormatCode {
 			case TextFormatCode: